% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviationThroughTime.R
\name{deviationThroughTime}
\alias{deviationThroughTime}
\title{deviationThroughTime}
\usage{
deviationThroughTime(variableDirectory, timePeriod)
}
\arguments{
\item{variableDirectory}{A directory containing at least two time slice rasters for a given climate variable.}

\item{timePeriod}{A number, in years, representing time elapsed between climate variable raster slices.}
}
\value{
A raster showing the geographic distribution of standard deviations through time in a particular climate variable.
}
\description{
A function that reads time-slice rasters of data for a given climate (typically processed data from a climate model run, suceh as the results of an analysis using PaleoView (Fordham, *et al*. 2017, Ecography)) in a given directory and calculates standard deviation in that climate variable through time.
}
\examples{
\dontrun{
precipDeviation <- deviationThroughTime("precipfiles/", 20000);
}

}
