% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absLatitudinalMean.R
\name{absLatitudinalMean}
\alias{absLatitudinalMean}
\title{absLatitudinalMean}
\usage{
absLatitudinalMean(rasterForCalculation)
}
\arguments{
\item{rasterForCalculation}{A raster that contains data for plotting according to latitudinal value}
}
\value{
A vector of mean raster values for each absolute value of latitude.
}
\description{
A function to calculate mean values of a raster at the absolute value of latitude, at the resolution of a given raster layer.
}
\examples{

data(precipDeviation);
precipStability <- 1/precipDeviation;
alm <- absLatitudinalMean(rasterForCalculation = precipStability);
plot(alm, main = "Precipitation Stability by Absolute Latitude",
ylab = "Relative Stability", type = "l");

}
\references{
Owens, HL, and RP Guralnick. *Submitted*. *Biodiversity Informatics*.
}
\seealso{
\code{\link{latitudinalMean}} for calculating mean values of rasters for all latitudinal bands.
}
\keyword{manip}
